Emails working. ToDo: Templates are ugly as hell!
