+++ templates/ResetPasswordTemplate.html
@@ -1,7 +1,6 @@
 <!doctype html>
-<html>
+<html lang="en">
 <head>
-  <title>{{ .Subject }}</title>
   <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1">
   <base target="_blank">
@@ -73,13 +72,14 @@
     }
     @media screen and (max-width: 600px) {
       .wrap {
-        max-width: auto;
+        max-width: fit-content;
       }
       .gutter {
         padding: 10px;
       }
     }
   </style>
+  <title>Reset Password</title>
 </head>
 <body style="background-color: #F0F1F3;font-family: 'Helvetica Neue', 'Segoe UI', Helvetica, sans-serif;font-size: 15px;line-height: 26px;margin: 0;color: #444;">
 <div class="gutter" style="padding: 30px;">&nbsp;</div>
@@ -88,9 +88,9 @@
   <h2>Reset Password</h2>
   <p>To reset the password, enter the following code on the password reset site</p>
   <div class="modal-content">
-    <span>gnsfjfwe</span>
+    <span>{{ .Code }}</span>
   </div>
-  <h3>If you accidentally closed the side: <a href="tas.technulgy.com/pwdReset">Password Reset</a></h3>
+  <h3><a href="tas.technulgy.com/#/resetPassword">Password Reset</a></h3>
   <h3>DO NOT SHARE THIS CODE!</h3>
   <h4>On issues, contact side admin: <a href="mailto:admin@technulgy.com">admin@technulgy.com</a> </h4>
 </div>
